Внесены изменения по замечаниям код-ревью. Поправлен баг с отображением таббара и навбара. Убраны лишние товары из DataSource. Переименованы контроллеры магазина и корзины. Добавлена документация на стартовый контроллер.
