+++ UiLuxury_TeamProject/Resources/Storyboards/Main.storyboard
@@ -2,7 +2,6 @@
 <document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="22155" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="POR-UJ-85G">
     <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
-        <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22131"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
@@ -25,10 +24,10 @@
             </objects>
             <point key="canvasLocation" x="-2609" y="858"/>
         </scene>
-        <!--Shop List View Controller-->
+        <!--Shop View Controller-->
         <scene sceneID="FNi-g1-BS6">
             <objects>
-                <tableViewController id="QRp-L8-TgJ" customClass="ShopListViewController" customModule="UiLuxury_TeamProject" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="QRp-L8-TgJ" customClass="ShopViewController" customModule="UiLuxury_TeamProject" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" id="ycJ-6J-ZNC">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -316,10 +315,10 @@
             </objects>
             <point key="canvasLocation" x="1896.8" y="858.62068965517244"/>
         </scene>
-        <!--Basket List View Controller-->
+        <!--Basket View Controller-->
         <scene sceneID="TGt-T0-igx">
             <objects>
-                <tableViewController id="AF3-cN-Ij7" customClass="BasketListViewController" customModule="UiLuxury_TeamProject" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="AF3-cN-Ij7" customClass="BasketViewController" customModule="UiLuxury_TeamProject" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" id="xR0-GD-wsl">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
